+++ caeroc/gui/runtime.py
@@ -1,17 +1,28 @@
-# -*- coding: utf-8 -*-
+"""Provides class which connects formulae and data to UI elements through
+Qt signals and slots.
+
+"""
 import sys
-from caeroc import formulae
 try:
+    from PyQt5 import QtCore
+    from PyQt5.QtCore import pyqtSlot as Slot
+    from PyQt5.QtWidgets import QDialog, QApplication
+    from PyQt5.QtGui import QStandardItemModel
+    use_pyqt = True
+except ImportError:
     from PySide import QtCore
     from PySide.QtCore import Slot
     from PySide.QtGui import QDialog, QStandardItemModel, QApplication
-    from caeroc.gui.base_pyside import Ui_CalcDialog
-except ImportError:
-    from PyQt5 import QtCore
-    from PyQt5.QtCore import pyqtSlot as Slot
-    from PyQt5.QtWidgets import QDialog
-    from PyQt5.QtGui import QStandardItemModel
-    from caeroc.gui.base_pyqt import Ui_CalcDialog
+    use_pyqt = False
+
+from .. import formulae
+from ..logger import logger
+if use_pyqt:
+    from .base_pyqt import Ui_CalcDialog
+    logger.debug('Using PyQt5 backend.')
+else:
+    from .base_pyside import Ui_CalcDialog
+    logger.debug('Using PySide backend.')
 
 
 class CalcDialog(QDialog):
@@ -20,29 +31,33 @@ class CalcDialog(QDialog):
     TODO: Error handling
     """
     def __init__(self, parent=None):
-        QDialog.__init__(self, parent)
+        """Initialize scalar values which keep track of different options in
+        the GUI and setup the application user interface.
+
+        """
+        super(CalcDialog, self).__init__(parent)
 
         self.ui = Ui_CalcDialog()
         self.ui.setupUi(self)
         # ----------Input------------
         self.key1 = self.ui.qcb1_input
         self.key2 = self.ui.qcb2_input
-        self._default_values()
-        # ---------Output------------
-        self.table = self.ui.qtw_output
-        self._setupModel()
-
-    def _default_values(self):
-        """Initialize scalar values which keep track of different options in
-        the GUI.
-
-        """
-        self.input1 = 0.5
-        self.input2 = 0.0
+        self.input1 = 1.0
+        self.input2 = 1.0
         self.gamma = 1.4
         self.autocalc = False
         self.mode = None
         self.on_qrb1_isen_pressed()
+        # ---------Output------------
+        self.table = self.ui.qtw_output
+        self._setupModel()
+
+        # -----Connect Signals-------
+        self.autocalc = False
+        if use_pyqt:
+            self.ui.qcb_autocalc.setChecked(self.autocalc)
+            self.ui.qcb_autocalc.stateChanged.connect(
+                self.on_qcb_autocalc_stateChanged)
 
     def _setupModel(self):
         """Sets up the model or the table where the output is displayed."""
@@ -64,9 +79,16 @@ class CalcDialog(QDialog):
             self.key1_to_attr = key1_to_attr
             self.key2_to_attr = key2_to_attr
             self._set_attr_to_key(extra_attr_to_key)
-            self.key1.addItems(self.key1_to_attr.keys())
-            self.key2.addItems(self.key2_to_attr.keys())
-            print('MODE: {}'.format(self.mode.__doc__))
+
+            keys1 = self.key1_to_attr.keys()
+            keys2 = self.key2_to_attr.keys()
+            if not use_pyqt:
+                keys1 = list(keys1)
+                keys2 = list(keys2)
+
+            self.key1.addItems(keys1)
+            self.key2.addItems(keys2)
+            logger.debug('MODE: {}'.format(self.mode.__doc__))
 
     def _set_attr_to_key(self, extra=None):
         self.attr_to_key = {v: k for k, v in self.key1_to_attr.items()}
@@ -75,7 +97,7 @@ class CalcDialog(QDialog):
 
     @Slot()
     def on_qrb1_isen_pressed(self):
-        """Run when radio button is pressed."""
+        """Run when radio button Isentropic is pressed."""
         self.Mode = formulae.isentropic.Isentropic
         key1_to_attr = {'M': 'M',
                         u'p/p\u2080': 'p_p0',
@@ -92,7 +114,7 @@ class CalcDialog(QDialog):
 
     @Slot()
     def on_qrb2_expa_pressed(self):
-        """Run when radio button is pressed."""
+        """Run when radio button Expansion is pressed."""
         self.Mode = formulae.isentropic.Expansion
         key1_to_attr = {u'θ (deg)': 'theta_deg',
                         u'θ (rad)': 'theta_rad'}
@@ -107,9 +129,29 @@ class CalcDialog(QDialog):
         self._set_mode_with_keys_and_attrs(key1_to_attr, key2_to_attr,
                                            extra_attr_to_key)
 
+    @Slot()
+    def on_qrb3_norm_pressed(self):
+        """Run when radio button Normal Shock is pressed."""
+        logger.warn('MODE:Normal Shock not implemented')
+
+    @Slot()
+    def on_qrb4_obli_pressed(self):
+        """Run when radio button Oblique Shock is pressed."""
+        logger.warn('MODE:Oblique Shock not implemented')
+
+    @Slot()
+    def on_qrb5_fann_pressed(self):
+        """Run when radio button Fanno Flow is pressed."""
+        logger.warn('MODE:Fanno Flow not implemented')
+
+    @Slot()
+    def on_qrb6_rayl_pressed(self):
+        """Run when radio button Rayleigh Flow is pressed."""
+        logger.warn('MODE:Rayleigh Flow not implemented')
+
     @Slot(float)
     def on_qdsb1_input_valueChanged(self, value):
-        """Run when input value in the QDoubleSpinBox changes."""
+        """Run when input value in the 1st input QDoubleSpinBox changes."""
         self.input1 = value
         if self.autocalc:
             self.on_qpb_calculate_released()
@@ -132,13 +174,16 @@ class CalcDialog(QDialog):
 
     @Slot()
     def on_qcb_autocalc_stateChanged(self):
-        """Run when the QComboBox state changes."""
-        self.autocalc = not self.autocalc
-        print('Auto calculate: ', self.autocalc)
+        """Run when the AutoCalculate QCheckBox state changes."""
+        self.autocalc = self.ui.qcb_autocalc.isChecked()
+        logger.debug('Auto calculate: {}'.format(self.autocalc))
 
     @Slot()
     def on_qpb_calculate_released(self):
-        """Run when the QPushButton is pressed and released."""
+        """Remove the rows and update the model with new data.
+        Runs when the QPushButton is pressed and released.
+
+        """
         attr1 = self.key1_to_attr[self.key1.currentText()]
         attr2 = self.key2_to_attr[self.key2.currentText()]
         if attr2 is None:
@@ -146,9 +191,9 @@ class CalcDialog(QDialog):
         else:
             kwargs = {attr1: self.input1, attr2: self.input2}
 
-        print(kwargs)
+        logger.info(kwargs)
         self.mode.calculate(**kwargs)
-        print(self.mode.data)
+        logger.info(self.mode.data)
 
         # ------ Fill table --------------
         self.model.removeRows(0,