+++ CHANGELOG.rst
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+0.0.2b2
+-------
+- logging/colorlog outputs
+- Fix autocalculate on PyQt5
+- Raise error for unusual exit on caeroc-test
+
 0.0.2b0
 -------
 - First beta release